@@ -4,13 +4,13 @@ version: 0.1.0+1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  path_provider: '>=1.4.0'
+  path_provider: '>=2.0.11'
 
   cupertino_icons: ^0.1.2
 
