@@ -1,11 +1,10 @@
 name: flutter_share
 description: Plugin for summoning a platform share sheet.
-version: 0.1.0
-author: LimeVista
+version: 1.0.0
 homepage: https://github.com/LimeVista/flutter_share
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
